metaobjects: example -> recipe
