+++ cookbook/recipes/metaobjects/ingredients/templates/skeleton/app/sections/SectionHero.tsx
@@ -53,8 +53,9 @@ export function SectionHero(props: SectionHeroFragment) {
               textDecoration: 'underline',
               marginTop: '1rem',
             }}
-            target={link?.target?.value === 'false' ? '_self' : '_blank'}
-            rel="noreferrer"
+            {...(link?.target?.value !== 'false'
+              ? {target: '_blank', rel: 'noreferrer'}
+              : {})}
           >
             {link?.text?.value}
           </Link>
