+++ cookbook/recipes/metaobjects/ingredients/templates/skeleton/app/sections/SectionStores.tsx
