+++ cookbook/recipes/metaobjects/ingredients/templates/skeleton/app/sections/Sections.tsx
