+++ cookbook/recipes/metaobjects/patches/README.md.47d06f.patch
@@ -1,0 +1,100 @@
+index c584e5370..927d99f1a 100644
+--- a/templates/skeleton/README.md
++++ b/templates/skeleton/README.md
+@@ -1,6 +1,8 @@
+-# Hydrogen template: Skeleton
++# Hydrogen template: Metaobjects as CMS
+ 
+-Hydrogen is Shopify’s stack for headless commerce. Hydrogen is designed to dovetail with [Remix](https://remix.run/), Shopify’s full stack web framework. This template contains a **minimal setup** of components, queries and tooling to get started with Hydrogen.
++This Hydrogen template demonstrates how to use Shopify Metaobjects as a content management system (CMS). Hydrogen is Shopify's stack for headless commerce, designed to work with [Remix](https://remix.run/), Shopify's full stack web framework.
++
++This template shows how to create a flexible, section-based content architecture using Shopify's native Metaobjects, allowing merchants to manage content directly from the Shopify admin without external CMS dependencies.
+ 
+ [Check out Hydrogen docs](https://shopify.dev/custom-storefronts/hydrogen)
+ [Get familiar with Remix](https://remix.run/docs/en/v1)
+@@ -16,18 +18,60 @@ Hydrogen is Shopify’s stack for headless commerce. Hydrogen is designed to dov
+ - Prettier
+ - GraphQL generator
+ - TypeScript and JavaScript flavors
+-- Minimal setup of components and routes
++- **Metaobjects-based CMS architecture**
++- **Dynamic section rendering system**
++- **Content management through Shopify admin**
++
++## Metaobjects Architecture
++
++This template implements a hierarchical content structure:
++
++```
++Route (Metaobject)
++  └── Sections (References)
++      ├── SectionHero
++      ├── SectionFeaturedProducts
++      ├── SectionFeaturedCollections
++      ├── SectionStoreProfile
++      └── SectionStoreGrid
++```
++
++### Key Features
++
++- **Route-based content**: Each route can have its own set of sections
++- **Reusable sections**: Create once, use across multiple routes
++- **Type-safe**: Full TypeScript support with generated types
++- **Merchant-friendly**: Content managed directly in Shopify admin
++- **Extensible**: Easy to add new section types
+ 
+ ## Getting started
+ 
+ **Requirements:**
+ 
+ - Node.js version 18.0.0 or higher
++- Shopify store with Metaobjects enabled
++- Metaobject definitions created in Shopify admin
+ 
+ ```bash
+ npm create @shopify/hydrogen@latest
+ ```
+ 
++## Setting up Metaobjects
++
++1. **Create Metaobject definitions** in your Shopify admin:
++   - Navigate to Settings → Custom data → Metaobjects
++   - Create definitions for Route, SectionHero, SectionFeaturedProducts, etc.
++   - See `guides/metaobjects/README.md` for detailed field configurations
++
++2. **Create content entries**:
++   - Add Route entries for pages you want to manage
++   - Create Section entries and link them to Routes
++   - Configure section content and references
++
++3. **Query and render**:
++   - Routes automatically query their associated sections
++   - Sections component handles dynamic rendering based on type
++
+ ## Building for production
+ 
+ ```bash
+@@ -40,6 +84,21 @@ npm run build
+ npm run dev
+ ```
+ 
++## Creating New Sections
++
++1. Define the Metaobject in Shopify admin
++2. Create a React component in `app/sections/`
++3. Add the GraphQL fragment for querying
++4. Register in the Sections component switch statement
++
++Example:
++```tsx
++export function SectionExample(props: SectionExampleFragment) {
++  const section = parseSection<...>(props);
++  return <section>...</section>;
++}
++```
++
+ ## Setup for using Customer Account API (`/account` section)
+ 
+-Follow step 1 and 2 of <https://shopify.dev/docs/custom-storefronts/building-with-the-customer-account-api/hydrogen#step-1-set-up-a-public-domain-for-local-development>
++Follow step 1 and 2 of <https://shopify.dev/docs/custom-storefronts/building-with-the-customer-account-api/hydrogen#step-1-set-up-a-public-domain-for-local-development>
+\ No newline at end of file
