+++ cookbook/recipes/metaobjects/patches/_index.tsx.82ab90.patch
@@ -1,0 +1,184 @@
+index 28102dbe6..27dbb95cf 100644
+--- a/templates/skeleton/app/routes/_index.tsx
++++ b/templates/skeleton/app/routes/_index.tsx
+@@ -1,19 +1,11 @@
+-import {
+-  Await,
+-  useLoaderData,
+-  Link,
+-} from 'react-router';
++import {useLoaderData} from 'react-router';
+ import type {Route} from './+types/_index';
+-import {Suspense} from 'react';
+-import {Image} from '@shopify/hydrogen';
+-import type {
+-  FeaturedCollectionFragment,
+-  RecommendedProductsQuery,
+-} from 'storefrontapi.generated';
+-import {ProductItem} from '~/components/ProductItem';
++
++// @description Add metaobject content imports
++import {ROUTE_CONTENT_QUERY, RouteContent} from '~/sections/RouteContent';
+ 
+ export const meta: Route.MetaFunction = () => {
+-  return [{title: 'Hydrogen | Home'}];
++  return [{title: 'Hydrogen Metaobject | Home'}];
+ };
+ 
+ export async function loader(args: Route.LoaderArgs) {
+@@ -31,14 +23,18 @@ export async function loader(args: Route.LoaderArgs) {
+  * needed to render the page. If it's unavailable, the whole page should 400 or 500 error.
+  */
+ async function loadCriticalData({context}: Route.LoaderArgs) {
+-  const [{collections}] = await Promise.all([
+-    context.storefront.query(FEATURED_COLLECTION_QUERY),
++  const {storefront} = context;
++
++  // @description Query the home route metaobject
++  const [{route}] = await Promise.all([
++    storefront.query(ROUTE_CONTENT_QUERY, {
++      variables: {handle: 'route-home'},
++      cache: storefront.CacheNone(),
++    }),
+     // Add other queries here, so that they are loaded in parallel
+   ]);
+ 
+-  return {
+-    featuredCollection: collections.nodes[0],
+-  };
++  return {route};
+ }
+ 
+ /**
+@@ -47,125 +43,17 @@ async function loadCriticalData({context}: Route.LoaderArgs) {
+  * Make sure to not throw any errors here, as it will cause the page to 500.
+  */
+ function loadDeferredData({context}: Route.LoaderArgs) {
+-  const recommendedProducts = context.storefront
+-    .query(RECOMMENDED_PRODUCTS_QUERY)
+-    .catch((error: Error) => {
+-      // Log query errors, but don't throw them so the page can still render
+-      console.error(error);
+-      return null;
+-    });
+-
+-  return {
+-    recommendedProducts,
+-  };
++  // No deferred data for this route
++  return {};
+ }
+ 
+ export default function Homepage() {
+-  const data = useLoaderData<typeof loader>();
++  const {route} = useLoaderData<typeof loader>();
++
+   return (
+     <div className="home">
+-      <FeaturedCollection collection={data.featuredCollection} />
+-      <RecommendedProducts products={data.recommendedProducts} />
++      {/* @description Render the route's content sections */}
++      <RouteContent route={route} />
+     </div>
+   );
+ }
+-
+-function FeaturedCollection({
+-  collection,
+-}: {
+-  collection: FeaturedCollectionFragment;
+-}) {
+-  if (!collection) return null;
+-  const image = collection?.image;
+-  return (
+-    <Link
+-      className="featured-collection"
+-      to={`/collections/${collection.handle}`}
+-    >
+-      {image && (
+-        <div className="featured-collection-image">
+-          <Image data={image} sizes="100vw" />
+-        </div>
+-      )}
+-      <h1>{collection.title}</h1>
+-    </Link>
+-  );
+-}
+-
+-function RecommendedProducts({
+-  products,
+-}: {
+-  products: Promise<RecommendedProductsQuery | null>;
+-}) {
+-  return (
+-    <div className="recommended-products">
+-      <h2>Recommended Products</h2>
+-      <Suspense fallback={<div>Loading...</div>}>
+-        <Await resolve={products}>
+-          {(response) => (
+-            <div className="recommended-products-grid">
+-              {response
+-                ? response.products.nodes.map((product) => (
+-                    <ProductItem key={product.id} product={product} />
+-                  ))
+-                : null}
+-            </div>
+-          )}
+-        </Await>
+-      </Suspense>
+-      <br />
+-    </div>
+-  );
+-}
+-
+-const FEATURED_COLLECTION_QUERY = `#graphql
+-  fragment FeaturedCollection on Collection {
+-    id
+-    title
+-    image {
+-      id
+-      url
+-      altText
+-      width
+-      height
+-    }
+-    handle
+-  }
+-  query FeaturedCollection($country: CountryCode, $language: LanguageCode)
+-    @inContext(country: $country, language: $language) {
+-    collections(first: 1, sortKey: UPDATED_AT, reverse: true) {
+-      nodes {
+-        ...FeaturedCollection
+-      }
+-    }
+-  }
+-` as const;
+-
+-const RECOMMENDED_PRODUCTS_QUERY = `#graphql
+-  fragment RecommendedProduct on Product {
+-    id
+-    title
+-    handle
+-    priceRange {
+-      minVariantPrice {
+-        amount
+-        currencyCode
+-      }
+-    }
+-    featuredImage {
+-      id
+-      url
+-      altText
+-      width
+-      height
+-    }
+-  }
+-  query RecommendedProducts ($country: CountryCode, $language: LanguageCode)
+-    @inContext(country: $country, language: $language) {
+-    products(first: 4, sortKey: UPDATED_AT, reverse: true) {
+-      nodes {
+-        ...RecommendedProduct
+-      }
+-    }
+-  }
+-` as const;
