+++ cookbook/recipes/metaobjects/patches/fragments.ts.5f86c7.patch
@@ -1,0 +1,29 @@
+index dc4426a9f..76d6a1d9c 100644
+--- a/templates/skeleton/app/lib/fragments.ts
++++ b/templates/skeleton/app/lib/fragments.ts
+@@ -231,3 +231,25 @@ export const FOOTER_QUERY = `#graphql
+   }
+   ${MENU_FRAGMENT}
+ ` as const;
++
++// @description Fragment for recommended products needed by ProductItem component
++export const RECOMMENDED_PRODUCT_FRAGMENT = `#graphql
++  fragment RecommendedProduct on Product {
++    id
++    title
++    handle
++    priceRange {
++      minVariantPrice {
++        amount
++        currencyCode
++      }
++    }
++    featuredImage {
++      id
++      url
++      altText
++      width
++      height
++    }
++  }
++` as const;
