+++ cookbook/recipes/metaobjects/patches/package.json.cfee98.patch
@@ -1,0 +1,14 @@
+index 0ee1599a1..7939e3b36 100644
+--- a/templates/skeleton/package.json
++++ b/templates/skeleton/package.json
+@@ -21,7 +21,9 @@
+     "react": "18.3.1",
+     "react-dom": "18.3.1",
+     "react-router": "7.8.2",
+-    "react-router-dom": "7.8.2"
++    "react-router-dom": "7.8.2",
++    "slate": "^0.101.4",
++    "slate-react": "^0.101.3"
+   },
+   "devDependencies": {
+     "@eslint/compat": "^1.2.5",
