--- examples/metaobjects/package.json
@@ -1,17 +1,0 @@
-{
-  "name": "example-metaobjects",
-  "private": true,
-  "prettier": "@shopify/prettier-config",
-  "scripts": {
-    "build": "shopify hydrogen build --codegen --diff",
-    "dev": "shopify hydrogen dev --codegen --diff",
-    "preview": "shopify hydrogen preview --build --diff",
-    "lint": "eslint --no-error-on-unmatched-pattern .",
-    "typecheck": "tsc --noEmit",
-    "codegen": "shopify hydrogen codegen"
-  },
-  "dependencies": {
-    "slate": "^0.101.4",
-    "slate-react": "^0.101.3"
-  }
-}
