--- examples/metaobjects/tsconfig.json
@@ -1,11 +1,0 @@
-{
-  "extends": "../../templates/skeleton/tsconfig.json",
-  "include": ["./**/*.d.ts", "./**/*.ts", "./**/*.tsx", "../../templates/skeleton/vite.config.ts"],
-  "compilerOptions": {
-    "baseUrl": ".",
-    "paths": {
-      "*": ["./*", "../../templates/skeleton/*"],
-      "~/*": ["app/*", "../../templates/skeleton/app/*"]
-    }
-  }
-}
